src/test/java/org/anne/aoc2024/Day13Test.java
package org.anne.aoc2024;

import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day13Test {

    List<String> input = Arrays.asList(
            "Button A: X+94, Y+34",
            "Button B: X+22, Y+67",
            "Prize: X=8400, Y=5400",

            "Button A: X+26, Y+66",
            "Button B: X+67, Y+21",
            "Prize: X=12748, Y=12176",

            "Button A: X+17, Y+86",
            "Button B: X+84, Y+37",
            "Prize: X=7870, Y=6450",

            "Button A: X+69, Y+23",
            "Button B: X+27, Y+71",
            "Prize: X=18641, Y=10279"
    );

    @Test
    void testPart1() {
        assertEquals(480, Day13.part1(input));
    }

    @Test
    void testPart2() {
        assertEquals(875318608908L, Day13.part2(input));
    }
}
